pyproject.toml
[tool.poetry]
name = "competitionaggregator"
version = "0.1.0"
description = ""
authors = ["den4ik2975 <98264297+den4ik2975@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.29"
aiosqlite = "^0.20.0"
loguru = "^0.7.2"
stackprinter = "^0.2.12"
pydantic-settings = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
